Added the "Phaseable" component to tiles.
The TileConstraint component now checks for the Phaseable component when
determining hits. If an object that requires the "TileConstraint" sets
its attemptPhase member to true (as the "PlatformControls" component now does
after a double-press of the DOWN_ARROW) then a hit won't be returned for
the phaseable platform, allowing an object to pass through the phaseable
tile.